refactor: Move the authentication tests in single file
