+++ tests/v0_4_0/test_vehicle_event.py
@@ -2,14 +2,12 @@ import html
 import json
 import uuid
 
-import jwt
 import pytest
 from flask import url_for
 
-from tests import utils
+from tests.utils import REGISTERED_DEVICE_ID, get_request, register_device
 
-from .conftest import REGISTRED_DEVICE_ID
-from .utils import generate_telemetry, get_request
+from .utils import generate_telemetry, get_timestamp
 
 
 def generate_payload(event):
@@ -17,7 +15,7 @@ def generate_payload(event):
     and trip_id"""
     payload = {
         'telemetry': generate_telemetry(),
-        'timestamp': utils.get_timestamp(),
+        'timestamp': get_timestamp(),
     }
     payload.update(event)
     return payload
@@ -31,8 +29,9 @@ def generate_payload(event):
         {'event_type': 'deregister', 'event_type_reason': 'missing'},
     ],
 )
-def test_valid_post(client, register_device, event_args):
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
+def test_valid_post(client, event_args):
+    register_device()
+    url = url_for('v0_4_0.vehicle_event', device_id=REGISTERED_DEVICE_ID)
     kwargs = get_request(generate_payload(event_args))
     response = client.post(
         url,
@@ -42,70 +41,12 @@ def test_valid_post(client, register_device, event_args):
     assert response.data == b''
 
 
-def test_incorrect_content_type(client, register_device):
-    # This is not handle by MDS 0.4.0, so it works with a bad Content-Type
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
-    event = {'event_type': 'register'}
-    kwargs = get_request(generate_payload(event))
-    kwargs['content_type'] = 'test/html'
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '201 CREATED'
-    assert response.data == b''
-
-
-def test_incorrect_authorization(client, register_device):
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
-    event = {'event_type': 'register'}
-    kwargs = get_request(generate_payload(event))
-    # With no auth at all
-    del kwargs['headers']['Authorization']
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '401 UNAUTHORIZED'
-    assert b'Please provide an Authorization' in response.data
-
-    # With Basic token
-    kwargs = get_request(generate_payload(event))
-    kwargs['headers']['Authorization'] = 'Basic dXNlcm5hbWU6cGFzc3dvcmQ='
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '401 UNAUTHORIZED'
-    assert b'Please provide a Bearer token' in response.data
-
-    # Provider_id is missing
-    kwargs = get_request(generate_payload(event))
-    token = jwt.encode({'key': 'value'}, 'secret')
-    kwargs['headers']['Authorization'] = 'Bearer %s' % token
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '401 UNAUTHORIZED'
-    assert b'Please provide a provider_id' in response.data
-
-    # Invalid JWT
-    kwargs = get_request(generate_payload(event))
-    kwargs['headers']['Authorization'] = 'Bearer bad_jwt'
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '401 UNAUTHORIZED'
-    assert b'Please provide a valid JWT' in response.data
-
-
-def test_trip_id(client, register_device):
+def test_trip_id(client):
     """should only be present for trip events"""
+    register_device()
 
     # trip_ip should be present
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
+    url = url_for('v0_4_0.vehicle_event', device_id=REGISTERED_DEVICE_ID)
     event = {'event_type': 'trip_start'}
     kwargs = get_request(generate_payload(event))
     response = client.post(
@@ -128,11 +69,12 @@ def test_trip_id(client, register_device):
     assert expected.encode() in response.data
 
 
-def test_event_type_reason(client, register_device):
+def test_event_type_reason(client):
     """required and allowed vaules depends on event_type"""
+    register_device()
 
     # event_type_reason shouldn't be present
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
+    url = url_for('v0_4_0.vehicle_event', device_id=REGISTERED_DEVICE_ID)
     event = {'event_type': 'register', 'event_type_reason': 'compliance'}
     kwargs = get_request(generate_payload(event))
     response = client.post(
@@ -167,8 +109,9 @@ def test_event_type_reason(client, register_device):
     assert expected.encode() in response.data
 
 
-def test_device_id(client, register_device):
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
+def test_device_id(client):
+    register_device()
+    url = url_for('v0_4_0.vehicle_event', device_id=REGISTERED_DEVICE_ID)
     event = {'event_type': 'register'}
     data = generate_payload(event)
     data['telemetry']['device_id'] = str(uuid.uuid4())
@@ -182,8 +125,10 @@ def test_device_id(client, register_device):
     assert expected.encode() in response.data
 
 
-def test_invalid_telemetry(client, register_device):
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
+def test_invalid_telemetry(client):
+    register_device()
+
+    url = url_for('v0_4_0.vehicle_event', device_id=REGISTERED_DEVICE_ID)
     event = {'event_type': 'register'}
     data = generate_payload(event)
     del data['telemetry']['gps']['lat']
@@ -197,8 +142,10 @@ def test_invalid_telemetry(client, register_device):
     assert expected.encode() in response.data
 
 
-def test_missing_required(client, register_device):
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
+def test_missing_required(client):
+    register_device()
+
+    url = url_for('v0_4_0.vehicle_event', device_id=REGISTERED_DEVICE_ID)
     event = {'event_type': 'register'}
     data = generate_payload(event)
     del data['timestamp']
@@ -212,8 +159,10 @@ def test_missing_required(client, register_device):
     assert expected.encode() in response.data
 
 
-def test_wrong_type(client, register_device):
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
+def test_wrong_type(client):
+    register_device()
+
+    url = url_for('v0_4_0.vehicle_event', device_id=REGISTERED_DEVICE_ID)
     event = {'event_type': 'register'}
     data = generate_payload(event)
     data['telemetry'] = 346
@@ -227,8 +176,10 @@ def test_wrong_type(client, register_device):
     assert expected.encode() in response.data
 
 
-def test_unknown_field(client, register_device):
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
+def test_unknown_field(client):
+    register_device()
+
+    url = url_for('v0_4_0.vehicle_event', device_id=REGISTERED_DEVICE_ID)
     event = {'event_type': 'register'}
     data = generate_payload(event)
     data['unknown_field'] = 'nope'
@@ -242,8 +193,8 @@ def test_unknown_field(client, register_device):
     assert expected.encode() in response.data
 
 
-def test_unregistred_device(client):
-    url = url_for('v0_4_0.vehicle_event', device_id=REGISTRED_DEVICE_ID)
+def test_unregistered_device(client):
+    url = url_for('v0_4_0.vehicle_event', device_id=REGISTERED_DEVICE_ID)
     event = {'event_type': 'register'}
     kwargs = get_request(generate_payload(event))
     response = client.post(
