+++ tests/v0_4_0/test_vehicle_register.py
@@ -3,13 +3,10 @@ import json
 import random
 import uuid
 
-import jwt
 from flask import url_for
 
 from tests import utils
-
-from .conftest import REGISTRED_DEVICE_ID
-from .utils import get_request
+from tests.utils import REGISTERED_DEVICE_ID, get_request, register_device
 
 PROPULSION_TYPE = [
     'combustion',
@@ -57,10 +54,12 @@ def test_post(client):
     assert response.data == b''
 
 
-def test_already_registred(client, register_device):
+def test_already_registred(client):
+    register_device()
+
     data = generate_payload()
     url = url_for('v0_4_0.vehicle_register')
-    data['device_id'] = REGISTRED_DEVICE_ID
+    data['device_id'] = REGISTERED_DEVICE_ID
     kwargs = get_request(data)
     response = client.post(
         url,
@@ -68,77 +67,6 @@ def test_already_registred(client, register_device):
     )
     assert response.status == '409 CONFLICT'
     assert b'' in response.data
-
-
-def test_incorrect_content_type(client):
-    # This is not handle by MDS 0.4.0, so it works with a bad Content-Type
-    url = url_for('v0_4_0.vehicle_register')
-    kwargs = get_request(generate_payload())
-    kwargs['content_type'] = 'test/html'
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '201 CREATED'
-    assert response.data == b''
-
-
-def test_incorrect_authorization(client):
-    url = url_for('v0_4_0.vehicle_register')
-    # With no auth at all
-    kwargs = get_request(generate_payload())
-    del kwargs['headers']['Authorization']
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '401 UNAUTHORIZED'
-    assert b'Please provide an Authorization' in response.data
-
-    # With Basic token
-    kwargs = get_request(generate_payload())
-    kwargs['headers']['Authorization'] = 'Basic dXNlcm5hbWU6cGFzc3dvcmQ='
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '401 UNAUTHORIZED'
-    assert b'Please provide a Bearer token' in response.data
-
-    # Provider_id is missing
-    kwargs = get_request(generate_payload())
-    token = jwt.encode({'key': 'value'}, 'secret')
-    kwargs['headers']['Authorization'] = 'Bearer %s' % token
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '401 UNAUTHORIZED'
-    assert b'Please provide a provider_id' in response.data
-
-    # Invalid JWT
-    kwargs = get_request(generate_payload())
-    kwargs['headers']['Authorization'] = 'Bearer bad_jwt'
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '401 UNAUTHORIZED'
-    assert b'Please provide a valid JWT' in response.data
-
-
-def test_missing_required(client):
-    url = url_for('v0_4_0.vehicle_register')
-    data = generate_payload()
-    del data['device_id']
-    kwargs = get_request(data)
-    response = client.post(
-        url,
-        **kwargs,
-    )
-    assert response.status == '400 BAD REQUEST'
-    expected = html.escape(json.dumps({'missing_param': ['device_id']}))
-    assert expected.encode() in response.data
 
 
 def test_wrong_type(client):
